Dont run initialisers if csv dialog cancelled
If the user cancelled the file dialog to select the unit.csv file then
the initialisers were still running. It should instead show the same
error message as if they had clicked no when asked if they had a
unit.csv file.

Additionally we reviewed the default directory for that file dialog
and decided to leave it as the current directory as the the home
directory on windows may be more confusing to many users.