Declare splash screen in jar manifest
Now that we are using an executable jar for Romeo we declare the splash
screen in the manifest instead of on the command line in the shell
scripts. 

Note that this is for the initial splash screen. Our own Swing one with
the progressor still takes over as soon as the application has started
running.