Standardise service method names to getX
The service methods used a mixture of the term 'get' and 'load' and
these have now been standardised to all use 'get'. Historically 'load'
was more descriptive for certain of the methods but now most of the
services load all the records and cache them internally anyway. Its more
intuitive to standardise on get.