Eliminate context.xml and drop some Spring dependencies
Modify RomeoContext to build the various objects itself instead of
using a Spring context. Remove the context.xml file and those
Spring dependencies we no longer need. We must still retain
spring-core for the time being as we still make a lot of use of the
ClassPathResource class that it provides.

n.b. the loss of context.xml means the expected order of csv columns
will not be tweakable via that file. It will be hardcoded in this
commit, but we need to look to making it user tweakable again lest
UC column order be changed in future.