Replace RomeoContext.createBattleCalculator with BattleCalculatorFactory
Previously BattlePanel would just using the Spring Context as a
service locator and get one from there so we could specify the
implementation (actually we only have one so far, but I'd still
prefer clients to use it via an interface) and its dependencies
there. With the move away from having a Spring Context, lets instead
give BattlePanel (and anything else that might need to create
instances of BattleCalculator) an object that can keep a reference
to the things it needs (currently just an xfactor compiler, but
BattlePanel doesn't need to know that!) and they can use it to
create their calculator instances on demand without having to know
which impl to use and without having to marshal the dependencies for
it themselves.