Replace RomeoContext.createWorldForm with a WorldFormFactory
The factory can be injected into things that need to create a
WorldForm so they don't need to reference the RomeoContext. The
factory object can wrap the references to the dependencies needed
for creating a WorldForm so things that need to create a WorldForm
don't need to worry about how, they just get the factory to do it
for them.