RomeoContext is banished
The RomeoContext class is no longer working as a 'context' or
service locator. Its more of a factory and does almost all its
work in the constructor. We can move that work to a static method
in the Romeo class itself and it can return the Romeo instance
object directly.

Also, how often do you get to say "from forth the fatal loins" in
code? ;-)