Run the database initialisers as early as possible
Recent refactoring broke startup when there is no database as
something that was constructed as part of this process attempted
to read from the db before it had been initialised.

Create a DatabaseInitialiser helpr class and move runInitialisers
and updateTriggerField to that, and invoke it directly inside the
static fromForthTheFatalLoins method as soon as we have constructed
the minimal set of service objects necessary to satisfy the
dependencies of all the initialisers. Construct all the rest of the
objects after this point.