Refactoring of form creation
Currently the code to create the components in a RomeoForm based on
the field definitions is in the RomeoForm class itself and so for
those field types that need a service we have to call a setter on
the form to provide that reference. I'd like to move away from using
setters for basic initialisation, I think such things should be
passed to the constructor and made mandatory to be able to create
the object. Id also like to make the construction of a form a more
one-shot process.

To this end, lets try moving the construction logic (and associated
dependency references) into a builder object, call it R
omeoFormInitialiser, and we can create an instance of that and
inject it into the form constructors together with the logic and
record etc.

This means reworking a lot of the existing form's plumbing, and as
part of this let's split the world form and its logic back into two
distinct classes again. (Actually we could consider going the other
way and eliminating the form/logic dichotomy everywhere as it was
mainly done that way to facilitate declaring forms and their fields
in a spring context.xml which is something we no longer do, however
this idea is something for another commit).