Fix opening world from the map
Recent changes to WorldForm and related classes introduced a bug
when opening the world from the map. The listener is now casting the
record object to IWorld, but when invoked by a map click the object
is a WorldAndHistory (which doesn't implement IWorld, rather it
wraps an IWorld and an IHistory object). The WorldFormLogic knows
how to handle this but the call to its bind never happens because
of the exception.

Fix by having WorldFormFactory.newWorldForm take an Object for the
record instead of IWorld and don't cast the record object to IWorld
in WorldNavigatorRecordSelectionListener.newForm