Change from commons-logging to SLF4J
Instead of using commons-logging for Romeo's own logging let's try
using SLF4J. Update all the code that needs to be updated and use
the SLF4J Simple Logger implementation.

Some of our dependencies use commons-logging and will pull it in as
a transitive dependency, so continue to include it in the POM but
as a runtime dependency.